refactor(api): extrait les regles de limites de scopes
validateObjectValue et validateScopeLimits partent dans
src/middleware/scope-limits.ts. Ce sont 66 lignes de regles metier pures,
sans dependance a Zod ni a Hono, donc directement testables, et le fichier
de routes perd son role annexe de porteur des regles de limites.

Le decoupage complet de routes/ui.tsx n'est deliberement pas fait ici. 42
des 47 schemas ne servent qu'a un seul endpoint : le couplage est vertical,
a travers les trois couches, donc un decoupage horizontal en schemas,
routes et handlers irait contre le sens du couplage et ferait toucher trois
fichiers a chaque modification d'endpoint. Le decoupage vertical est le bon
mais coute cher, uiRoutes etant une instance OpenAPIHono unique et mutable.
Il merite sa propre PR.
